add Darwin to build
