+++ .github/workflows/push.yaml
@@ -26,11 +26,16 @@ jobs:
     with:
       GOARCH: arm64
     secrets: inherit
-  # darwin_amd64:
-  #   uses: ./.github/workflows/build-darwin.yaml
-  #   with:
-  #     GOARCH: amd64
-  #   secrets: inherit
+  darwin_amd64:
+    uses: ./.github/workflows/build-darwin.yaml
+    with:
+      GOARCH: amd64
+    secrets: inherit
+  darwin_arm64:
+    uses: ./.github/workflows/build-darwin.yaml
+    with:
+      GOARCH: arm64
+    secrets: inherit
   release:
     concurrency: release
     runs-on: ubuntu-latest
@@ -42,7 +47,8 @@ jobs:
       - linux_amd64
       - linux_arm
       - linux_arm64
-      # - darwin_amd64
+      - darwin_amd64
+      - darwin_arm64
     steps:
       - uses: actions/download-artifact@v3
         with:
@@ -56,9 +62,12 @@ jobs:
       - uses: actions/download-artifact@v3
         with:
           name: ${{ needs.linux_arm64.outputs.artifact }}
-      # - uses: actions/download-artifact@v3
-      #   with:
-      #     name: ${{ needs.darwin_amd64.outputs.artifact }}
+      - uses: actions/download-artifact@v3
+        with:
+          name: ${{ needs.darwin_amd64.outputs.artifact }}
+      - uses: actions/download-artifact@v3
+        with:
+          name: ${{ needs.darwin_arm64.outputs.artifact }}
       - name: Release
         uses: softprops/action-gh-release@v1
         with:
@@ -67,7 +76,8 @@ jobs:
             ${{ needs.linux_amd64.outputs.artifact }}
             ${{ needs.linux_arm.outputs.artifact }}
             ${{ needs.linux_arm64.outputs.artifact }}
-          # ${{ needs.darwin_amd64.outputs.artifact }}
+            ${{ needs.darwin_amd64.outputs.artifact }}
+            ${{ needs.darwin_arm64.outputs.artifact }}
       - name: Add to pkg.go.dev
         # https://pkg.go.dev/about#adding-a-package
         run: |
