+++ cli/lib/lib_darwin.go
@@ -1,0 +1,36 @@
+package lib
+
+import (
+	"context"
+
+	"github.com/spf13/cobra"
+
+	"github.com/fornellas/resonance/host"
+)
+
+func Reset() {
+	resetCommon()
+}
+
+func GetHost(ctx context.Context) (host.Host, error) {
+	var hst host.Host
+	var err error
+
+	hst, err = host.NewSshAuthority(ctx, hostname)
+	if err != nil {
+		return nil, err
+	}
+
+	if sudo {
+		hst, err = host.NewSudo(ctx, hst)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	return hst, nil
+}
+
+func AddHostFlags(cmd *cobra.Command) {
+	addHostFlagsCommon(cmd)
+}
